Repaired warnings

+++ ZPG_Projekt/DynamicRotate.h
@@ -10,7 +10,7 @@ public:
     DynamicRotate(float angle, const glm::vec3& axis) : angle(angle), axis(axis) {}
 
     void apply(glm::mat4& modelMatrix) const override {
-        float newAngle = glfwGetTime() * angle;
+        float newAngle = (float)glfwGetTime() * angle;
         modelMatrix = glm::rotate(modelMatrix, glm::radians(newAngle), axis);
     }
 };
